requirements.txt
PyQt5
requests
pygame
python-dotenv
numpy
sounddevice
scipy
pydub
scikit-learn